# insis_2020_2021_1161393_1161491/p3/resources/ProposalCommandOpenApi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "OpenAPI definition",
		"version": "v0"
	},
	"servers": [
		{
			"url": "http://localhost:8081",
			"description": "Generated server url"
		}
	],
	"paths": {
		"/api/proposals/{id}": {
			"put": {
				"tags": [
					"proposal-controller-impl"
				],
				"summary": "Update a proposal.",
				"operationId": "updateProposal",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id of proposal to be updated",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateProposalDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"404": {
						"description": "Invalid id supplied"
					},
					"200": {
						"description": "Proposal updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Proposal"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"delete": {
				"tags": [
					"proposal-controller-impl"
				],
				"summary": "Delete a proposal.",
				"operationId": "deleteProposal",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id of proposal to be deleted",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"404": {
						"description": "Invalid id supplied"
					},
					"200": {
						"description": "Proposal deleted"
					},
					"500": {
						"description": "Internal server error"
					}
				}
			},
			"patch": {
				"tags": [
					"proposal-controller-impl"
				],
				"summary": "Update a proposal status.",
				"operationId": "updateProposalStatus",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id of proposal to be updated",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "string",
								"description": "The new proposal status"
							}
						}
					},
					"required": true
				},
				"responses": {
					"404": {
						"description": "Invalid id supplied"
					},
					"200": {
						"description": "Proposal updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Proposal"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
...
		"/api/proposals": {
			"post": {
				"tags": [
					"proposal-controller-impl"
				],
				"summary": "Create a proposal.",
				"operationId": "createProposal",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProposalDTO"
							}
						}
					},
					"required": true
				},
				"responses": {
					"201": {
						"description": "Proposal created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Proposal"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error"
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"UpdateProposalDTO": {
				"required": [
					"goals",
					"problem",
					"proponentId",
					"proposalId",
					"status",
					"title"
				],
				"type": "object",
				"properties": {
					"proposalId": {
						"maxLength": 30,
						"type": "string",
						"description": "Proposal id"
					},
					"title": {
						"maxLength": 20,
						"type": "string",
						"description": "Proposal title"
					},
					"problem": {
						"maxLength": 20,
						"type": "string",
						"description": "Proposal description"
					},
					"goals": {
						"maxLength": 20,
						"type": "string",
						"description": "Proposal goals"
					},
					"technologies": {
						"type": "array",
						"description": "List of proposal technologies",
						"items": {
							"type": "string",
							"description": "List of proposal technologies"
						}
					},
					"keywords": {
						"type": "array",
						"description": "List of proposal keywords",
						"items": {
							"type": "string",
							"description": "List of proposal keywords"
						}
					},
					"approaches": {
						"type": "array",
						"description": "List of proposal approaches",
						"items": {
							"type": "string",
							"description": "List of proposal approaches"
						}
					},
					"proponentId": {
						"maxLength": 30,
						"type": "string",
						"description": "Proponent id"
					},
					"comment": {
						"maxLength": 20,
						"type": "string",
						"description": "RUC comment"
					},
					"status": {
						"maxLength": 20,
						"type": "string",
						"description": "Proposal status"
					}
				}
			},
			"Proposal": {
				"type": "object",
				"properties": {
					"proposalId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"problem": {
						"type": "string"
					},
					"goals": {
						"type": "string"
					},
					"technologies": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"keywords": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"approaches": {
						"type": "array",
						"items": {
							"type": "string"
						}
					},
					"proponentId": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					},
					"proposalStatus": {
						"type": "string",
						"enum": [
							"ANALYSIS",
							"REJECTED",
							"CANCELED",
							"ACCEPTED",
							"ERROR"
						]
					},
					"proposalTimestamp": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
			"CreateProposalDTO": {
				"required": [
					"goals",
					"problem",
					"proponentId",
					"title"
				],
				"type": "object",
				"properties": {
					"title": {
						"maxLength": 20,
						"type": "string",
						"description": "Proposal title"
					},
					"problem": {
						"maxLength": 20,
						"type": "string",
						"description": "Proposal description"
					},
					"goals": {
						"maxLength": 20,
						"type": "string",
						"description": "Proposal goals"
					},
					"technologies": {
						"type": "array",
						"description": "List of proposal technologies",
						"items": {
							"type": "string",
							"description": "List of proposal technologies"
						}
					},
					"keywords": {
						"type": "array",
						"description": "List of proposal keywords",
						"items": {
							"type": "string",
							"description": "List of proposal keywords"
						}
					},
					"approaches": {
						"type": "array",
						"description": "List of proposal approaches",
						"items": {
							"type": "string",
							"description": "List of proposal approaches"
						}
					},
					"proponentId": {
						"maxLength": 30,
						"type": "string",
						"description": "Proponent id"
					}
				}
			},
...
		}
	}
}